Fixed linters
