+++ src/content/enemylaser.js
@@ -1,0 +1,7 @@
+import Entity from './entity';
+
+export default class EnemyLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'LaserEnemy');
+  }
+}
