+++ src/content/player.js
@@ -1,5 +1,4 @@
-/* eslint-disable max-classes-per-file */
-import { Entity } from './entity';
+import Entity from './entity';
 
 class Player extends Entity {
   constructor(scene, x, y) {
@@ -8,10 +7,4 @@ class Player extends Entity {
   }
 }
 
-class PlayerLaser extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'LaserPlayer');
-  }
-}
-
-export { Player, PlayerLaser };
+export default Player;
