+++ src/content/playerlaser.js
@@ -1,0 +1,9 @@
+import Entity from './entity';
+
+class PlayerLaser extends Entity {
+  constructor(scene, x, y) {
+    super(scene, x, y, 'LaserPlayer');
+  }
+}
+
+export default PlayerLaser;
